refactor(payments): split PaymentService, add webhook secret tests, extend audit logging
Split the 1159-line PaymentService into focused services, following
the single-responsibility boundaries already implied by its
constructor: RefundService (refund transaction logic), ScheduleService
(recurring payment schedules), and PaymentWebhookService (payment
gateway webhook handling). PaymentService keeps core payment CRUD,
Stellar/escrow flows, and analytics. Controllers and the payment
module are rewired accordingly; existing tests were split/moved to
match the new service boundaries with no loss of coverage.

Added missing test coverage for the payment gateway webhook secret
check (valid/invalid/missing/empty secret header), which previously
had no regression protection - a plain string comparison that could
be silently broken without any test catching it.

Extended the existing @AuditLog audit trail to cover payment
operations that were previously invisible to it: Stellar rent
payments, escrow create/release/refund, retry-failed, reconciliation,
and the payment gateway webhook handler. Added the missing
AuditAction.ESCROW_REFUNDED action used by the escrow refund flow.
This plugs directly into the existing audit_logs table, retention
policy, and admin query/stats endpoints - no new logging
infrastructure was needed.

Closes #1366
Closes #1389
Closes #1376
