+++ backend/src/modules/audit/entities/audit-log.entity.ts
@@ -26,6 +26,7 @@ export enum AuditAction {
   PAYMENT_REFUNDED = 'PAYMENT_REFUNDED',
   ESCROW_CREATED = 'ESCROW_CREATED',
   ESCROW_RELEASED = 'ESCROW_RELEASED',
+  ESCROW_REFUNDED = 'ESCROW_REFUNDED',
   ESCROW_DISPUTED = 'ESCROW_DISPUTED',
   // Blockchain operations
   BLOCKCHAIN_TX_SUBMITTED = 'BLOCKCHAIN_TX_SUBMITTED',
