+++ backend/src/modules/payments/payment.module.ts
@@ -1,6 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PaymentService } from './payment.service';
+import { RefundService } from './refund.service';
+import { ScheduleService } from './schedule.service';
+import { PaymentWebhookService } from './payment-webhook.service';
 import {
   PaymentController,
   AgreementPaymentController,
@@ -38,7 +41,14 @@ import { AuditModule } from '../audit/audit.module';
     PaymentWebhookController,
     AdminRefundsController,
   ],
-  providers: [PaymentService, PaymentGatewayService, AdminRefundsService],
+  providers: [
+    PaymentService,
+    RefundService,
+    ScheduleService,
+    PaymentWebhookService,
+    PaymentGatewayService,
+    AdminRefundsService,
+  ],
   exports: [PaymentService, PaymentGatewayService],
 })
 export class PaymentModule {}
