+++ backend/src/modules/payments/payment.service.spec.ts
@@ -1,22 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { BadRequestException, NotFoundException } from '@nestjs/common';
-import { DataSource, Repository } from 'typeorm';
+import { BadRequestException } from '@nestjs/common';
+import { Repository } from 'typeorm';
 import { PaymentService } from './payment.service';
 import { Payment } from './entities/payment.entity';
 import { PaymentMethod } from './entities/payment-method.entity';
-import {
-  PaymentSchedule,
-  PaymentScheduleStatus,
-} from './entities/payment-schedule.entity';
 import { PaymentGatewayService } from './payment-gateway.service';
 import { NotificationsService } from '../notifications/notifications.service';
 import { PaymentStatus } from './entities/payment.entity';
 import { CreatePaymentRecordDto } from './dto/record-payment.dto';
-import { ProcessRefundDto } from './dto/process-refund.dto';
 import { CreatePaymentMethodDto } from './dto/create-payment-method.dto';
-import { CreatePaymentScheduleDto } from './dto/create-payment-schedule.dto';
-import { PaymentInterval } from './entities/payment-schedule.entity';
 import { PaymentProcessingService } from '../stellar/services/payment-processing.service';
 import { StellarService } from '../stellar/services/stellar.service';
 import * as StellarSdk from '@stellar/stellar-sdk';
@@ -41,14 +34,6 @@ const mockPaymentMethodRepository = () => ({
   save: jest.fn(),
   update: jest.fn(),
   remove: jest.fn(),
-  createQueryBuilder: jest.fn(),
-});
-
-const mockPaymentScheduleRepository = () => ({
-  findOne: jest.fn(),
-  find: jest.fn(),
-  create: jest.fn(),
-  save: jest.fn(),
   createQueryBuilder: jest.fn(),
 });
 
@@ -90,18 +75,6 @@ const mockIdempotencyService = {
   ),
 };
 
-// DataSource mock — transaction() runs the callback with a mock entity manager.
-const mockEntityManager = {
-  findOne: jest.fn(),
-  save: jest.fn(),
-};
-const mockDataSource = {
-  transaction: jest.fn(
-    (cb: (em: typeof mockEntityManager) => Promise<unknown>) =>
-      cb(mockEntityManager),
-  ),
-};
-
 const mockFraudHooksService = {
   onPaymentRecorded: jest.fn().mockResolvedValue(undefined),
 };
@@ -110,7 +83,6 @@ describe('PaymentService', () => {
   let service: PaymentService;
   let paymentRepository: Repository<Payment>;
   let paymentMethodRepository: Repository<PaymentMethod>;
-  let paymentScheduleRepository: Repository<PaymentSchedule>;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -123,10 +95,6 @@ describe('PaymentService', () => {
         {
           provide: getRepositoryToken(PaymentMethod),
           useFactory: mockPaymentMethodRepository,
-        },
-        {
-          provide: getRepositoryToken(PaymentSchedule),
-          useFactory: mockPaymentScheduleRepository,
         },
         {
           provide: PaymentGatewayService,
@@ -157,10 +125,6 @@ describe('PaymentService', () => {
           useValue: mockIdempotencyService,
         },
         {
-          provide: DataSource,
-          useValue: mockDataSource,
-        },
-        {
           provide: FraudHooksService,
           useValue: mockFraudHooksService,
         },
@@ -173,9 +137,6 @@ describe('PaymentService', () => {
     );
     paymentMethodRepository = module.get<Repository<PaymentMethod>>(
       getRepositoryToken(PaymentMethod),
-    );
-    paymentScheduleRepository = module.get<Repository<PaymentSchedule>>(
-      getRepositoryToken(PaymentSchedule),
     );
   });
 
@@ -443,142 +404,6 @@ describe('PaymentService', () => {
     });
   });
 
-  describe('processRefund', () => {
-    beforeEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('throws when payment is not found', async () => {
-      mockEntityManager.findOne.mockResolvedValue(null);
-
-      await expect(
-        service.processRefund(
-          'pay_1',
-          { amount: 10, reason: 'test' },
-          'user_1',
-        ),
-      ).rejects.toBeInstanceOf(NotFoundException);
-    });
-
-    it('processes refund successfully using pessimistic lock transaction', async () => {
-      const payment = {
-        id: 'pay_1',
-        userId: 'user_1',
-        status: PaymentStatus.COMPLETED,
-        amount: 100,
-        refundAmount: 0,
-        currency: 'NGN',
-        metadata: { chargeId: 'charge_1' },
-        user: {} as any,
-        agreementId: null,
-        transactionFee: 0,
-        netAmount: 100,
-        paymentMethod: null,
-        paymentMethodRelation: null,
-        paymentMethodRelationId: null,
-        receiptUrl: '',
-        referenceNumber: null,
-        processedAt: new Date(),
-        idempotencyKey: null,
-        refundStatus: 'none',
-        refundReason: null,
-        notes: null,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      } as unknown as Payment;
-
-      mockEntityManager.findOne.mockResolvedValue(payment);
-      mockPaymentGateway.processRefund.mockResolvedValue({
-        success: true,
-        refundId: 'refund_1',
-      });
-      mockEntityManager.save.mockResolvedValue({
-        ...payment,
-        status: PaymentStatus.REFUNDED,
-        refundAmount: 100,
-      });
-
-      const dto: ProcessRefundDto = { amount: 100, reason: 'test' };
-      const result = await service.processRefund('pay_1', dto, 'user_1');
-
-      // Verify pessimistic lock was requested.
-      expect(mockEntityManager.findOne).toHaveBeenCalledWith(
-        Payment,
-        expect.objectContaining({
-          lock: { mode: 'pessimistic_write' },
-        }),
-      );
-      expect(result.status).toBe(PaymentStatus.REFUNDED);
-      expect(mockNotificationsService.notify).toHaveBeenCalledWith(
-        'user_1',
-        'Refund processed',
-        expect.stringContaining('100'),
-        'PAYMENT_REFUNDED',
-      );
-    });
-
-    it('throws when refund amount exceeds available amount', async () => {
-      const payment = {
-        id: 'pay_1',
-        userId: 'user_1',
-        status: PaymentStatus.COMPLETED,
-        amount: 50,
-        refundAmount: 0,
-        metadata: { chargeId: 'charge_1' },
-      } as unknown as Payment;
-
-      mockEntityManager.findOne.mockResolvedValue(payment);
-
-      await expect(
-        service.processRefund(
-          'pay_1',
-          { amount: 100, reason: 'over' },
-          'user_1',
-        ),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('throws when charge id is missing', async () => {
-      const payment = {
-        id: 'pay_1',
-        userId: 'user_1',
-        status: PaymentStatus.COMPLETED,
-        amount: 100,
-        refundAmount: 0,
-        metadata: {},
-        currency: 'NGN',
-      } as unknown as Payment;
-
-      mockEntityManager.findOne.mockResolvedValue(payment);
-
-      await expect(
-        service.processRefund(
-          'pay_1',
-          { amount: 10, reason: 'test' },
-          'user_1',
-        ),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('prevents double-refund: second concurrent call sees updated refundAmount', async () => {
-      // Simulate the state after a first refund has already been applied.
-      const alreadyRefunded = {
-        id: 'pay_1',
-        userId: 'user_1',
-        status: PaymentStatus.REFUNDED,
-        amount: 100,
-        refundAmount: 100,
-        metadata: { chargeId: 'charge_1' },
-      } as unknown as Payment;
-
-      mockEntityManager.findOne.mockResolvedValue(alreadyRefunded);
-
-      await expect(
-        service.processRefund('pay_1', { amount: 1, reason: 'dup' }, 'user_1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-  });
-
   describe('createPaymentMethod', () => {
     it('creates payment method with encryption when sensitive metadata is provided', async () => {
       process.env.PAYMENT_METADATA_SECRET = 'test-secret';
@@ -614,92 +439,6 @@ describe('PaymentService', () => {
       expect(
         (createdPaymentMethod as Partial<PaymentMethod>)?.encryptedMetadata,
       ).toBeTruthy();
-    });
-  });
-
-  describe('createPaymentSchedule', () => {
-    it('creates payment schedule successfully', async () => {
-      const dto: CreatePaymentScheduleDto = {
-        agreementId: 'agreement_1',
-        paymentMethodId: '1',
-        amount: 100,
-        interval: PaymentInterval.MONTHLY,
-      };
-
-      (paymentMethodRepository.findOne as jest.Mock).mockResolvedValue({
-        id: 1,
-        userId: 'user_1',
-      });
-      const createScheduleMock = jest.spyOn(
-        paymentScheduleRepository,
-        'create',
-      );
-      createScheduleMock.mockImplementation(
-        (data: Partial<PaymentSchedule>) => data as PaymentSchedule,
-      );
-      (paymentScheduleRepository.save as jest.Mock).mockResolvedValue({
-        id: 'schedule_1',
-        ...dto,
-      });
-
-      const result = await service.createPaymentSchedule(dto, 'user_1');
-
-      expect(result.id).toBe('schedule_1');
-      const [createdSchedule] = createScheduleMock.mock.calls[0] ?? [];
-      expect((createdSchedule as Partial<PaymentSchedule>)?.status).toBe(
-        PaymentScheduleStatus.ACTIVE,
-      );
-    });
-  });
-
-  describe('runPaymentSchedule', () => {
-    it('throws when schedule is not active', async () => {
-      const schedule = {
-        id: 'schedule_1',
-        userId: 'user_1',
-        status: PaymentScheduleStatus.PAUSED,
-      } as PaymentSchedule;
-
-      (paymentScheduleRepository.findOne as jest.Mock).mockResolvedValue(
-        schedule,
-      );
-
-      await expect(
-        service.runPaymentSchedule('schedule_1', 'user_1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('fails schedule when payment method is missing', async () => {
-      const schedule = {
-        id: 'schedule_missing_method',
-        userId: 'user_1',
-        status: PaymentScheduleStatus.ACTIVE,
-        paymentMethodId: null,
-        retries: 0,
-        maxRetries: 3,
-        nextRunAt: new Date(),
-      } as unknown as PaymentSchedule;
-      (paymentScheduleRepository.findOne as jest.Mock).mockResolvedValue(
-        schedule,
-      );
-      (paymentScheduleRepository.save as jest.Mock).mockResolvedValue(schedule);
-
-      await expect(
-        service.runPaymentSchedule('schedule_missing_method', 'user_1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-
-      expect(paymentScheduleRepository.save).toHaveBeenCalledWith(
-        expect.objectContaining({
-          id: 'schedule_missing_method',
-          status: PaymentScheduleStatus.FAILED,
-        }),
-      );
-      expect(mockNotificationsService.notify).toHaveBeenCalledWith(
-        'user_1',
-        'Recurring payment failed',
-        expect.stringContaining('Payment method is missing'),
-        'PAYMENT_FAILED',
-      );
     });
   });
 
@@ -802,28 +541,6 @@ describe('PaymentService', () => {
 
       expect(result.updated).toBe(1);
       expect(mockStellarService.getEscrowById).toHaveBeenCalledWith(7);
-    });
-
-    it('updates payment from webhook event', async () => {
-      (paymentRepository.findOne as jest.Mock).mockResolvedValue({
-        id: 'pay_1',
-        status: PaymentStatus.PENDING,
-        metadata: {},
-      });
-      (paymentRepository.save as jest.Mock).mockResolvedValue({
-        id: 'pay_1',
-        status: PaymentStatus.COMPLETED,
-      });
-
-      const result = await service.handlePaymentGatewayWebhook({
-        eventType: 'payment.completed',
-        paymentId: 'pay_1',
-        status: 'completed',
-        transactionHash: 'tx_complete',
-      });
-
-      expect(result.processed).toBe(true);
-      expect((result.payment as Payment).status).toBe(PaymentStatus.COMPLETED);
     });
 
     it('retries failed payments that still have a payment method', async () => {
